requirements.txt
aniso8601==6.0.0
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
html2text==2018.1.9
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Paste==3.0.8
pycparser==2.19
PyJWT==1.4.2
pyodbc==4.0.26
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
visitor==0.1.3
waitress==1.3.0
Werkzeug==0.15.2

#Upload IMGEN
Pillow==2.2.1